refactor cli output helpers into separate files
