+++ internal/cli/client.go
@@ -1,0 +1,128 @@
+package cli
+
+import (
+	"context"
+	"fmt"
+	"io"
+	"net"
+	"net/http"
+	"net/url"
+	"os"
+	"path/filepath"
+	"strings"
+	"time"
+)
+
+func ResolveClientConfig(hostFlag string) (ClientConfig, error) {
+	hostFlag = strings.TrimSpace(hostFlag)
+	if hostFlag != "" {
+		baseURL, err := normalizeCLIHost("--host", hostFlag)
+		if err != nil {
+			return ClientConfig{}, commandExitError{Code: exitCodeUsage, Err: err}
+		}
+		return ClientConfig{
+			BaseURL:     baseURL,
+			Source:      "--host",
+			SourceValue: hostFlag,
+		}, nil
+	}
+
+	if envHost := strings.TrimSpace(os.Getenv("AGENT_COMPOSE_HOST")); envHost != "" {
+		baseURL, err := normalizeCLIHost("AGENT_COMPOSE_HOST", envHost)
+		if err != nil {
+			return ClientConfig{}, commandExitError{Code: exitCodeUsage, Err: err}
+		}
+		return ClientConfig{
+			BaseURL:     baseURL,
+			Source:      "AGENT_COMPOSE_HOST",
+			SourceValue: envHost,
+		}, nil
+	}
+
+	socketPath, err := resolveAgentComposeSocketForCLI(os.Getenv("AGENT_COMPOSE_SOCKET"))
+	if err != nil {
+		return ClientConfig{}, commandExitError{Code: exitCodeUsage, Err: err}
+	}
+	return ClientConfig{
+		BaseURL:       "http://agent-compose",
+		SocketPath:    socketPath,
+		Source:        "AGENT_COMPOSE_SOCKET",
+		SourceValue:   socketPath,
+		UseUnixSocket: true,
+	}, nil
+}
+
+func normalizeCLIHost(name, value string) (string, error) {
+	parsed, err := url.Parse(value)
+	if err != nil {
+		return "", fmt.Errorf("invalid %s %q: %w", name, value, err)
+	}
+	if parsed.Scheme != "http" && parsed.Scheme != "https" {
+		return "", fmt.Errorf("invalid %s %q: scheme must be http or https", name, value)
+	}
+	if parsed.Host == "" {
+		return "", fmt.Errorf("invalid %s %q: host is required", name, value)
+	}
+	return strings.TrimRight(value, "/"), nil
+}
+
+func resolveAgentComposeSocketForCLI(value string) (string, error) {
+	value = strings.TrimSpace(value)
+	if value == "" {
+		if runtimeDir := strings.TrimSpace(os.Getenv("XDG_RUNTIME_DIR")); runtimeDir != "" {
+			value = filepath.Join(runtimeDir, "agent-compose.sock")
+		} else {
+			value = filepath.Join(os.TempDir(), fmt.Sprintf("agent-compose-%d.sock", os.Getuid()))
+		}
+	}
+	if value == "" {
+		return "", fmt.Errorf("AGENT_COMPOSE_SOCKET is empty")
+	}
+	if strings.IndexByte(value, 0) >= 0 {
+		return "", fmt.Errorf("invalid AGENT_COMPOSE_SOCKET %q: path contains NUL byte", value)
+	}
+	resolved, err := filepath.Abs(value)
+	if err != nil {
+		return value, nil
+	}
+	return resolved, nil
+}
+
+func fetchDaemonVersion(ctx context.Context, clientConfig ClientConfig) ([]byte, error) {
+	client := newDaemonHTTPClient(clientConfig)
+	req, err := http.NewRequestWithContext(ctx, http.MethodGet, clientConfig.BaseURL+"/api/version", nil)
+	if err != nil {
+		return nil, fmt.Errorf("create daemon request for %s %q: %w", clientConfig.Source, clientConfig.SourceValue, err)
+	}
+	resp, err := client.Do(req)
+	if err != nil {
+		return nil, commandExitError{Code: exitCodeUnavailable, Err: fmt.Errorf("connect daemon via %s %q: %w", clientConfig.Source, clientConfig.SourceValue, err)}
+	}
+	defer func() {
+		_ = resp.Body.Close()
+	}()
+
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return nil, fmt.Errorf("read daemon response from %s %q: %w", clientConfig.Source, clientConfig.SourceValue, err)
+	}
+	if resp.StatusCode != http.StatusOK {
+		return nil, fmt.Errorf("daemon via %s %q returned HTTP %d: %s", clientConfig.Source, clientConfig.SourceValue, resp.StatusCode, strings.TrimSpace(string(body)))
+	}
+	return body, nil
+}
+
+func newDaemonHTTPClient(clientConfig ClientConfig) *http.Client {
+	transport := http.DefaultTransport.(*http.Transport).Clone()
+	if clientConfig.UseUnixSocket {
+		socketPath := clientConfig.SocketPath
+		transport.DialContext = func(ctx context.Context, network, addr string) (net.Conn, error) {
+			var dialer net.Dialer
+			return dialer.DialContext(ctx, "unix", socketPath)
+		}
+	}
+	return &http.Client{
+		Transport: transport,
+		Timeout:   10 * time.Minute,
+	}
+}
