+++ internal/cli/format.go
@@ -1,0 +1,172 @@
+package cli
+
+import (
+	"fmt"
+	"strings"
+	"time"
+
+	agentcomposev2 "agent-compose/proto/agentcompose/v2"
+)
+
+func manualRunClientRequestID(projectName, agentName, prompt string) string {
+	value := strings.TrimSpace(projectName) + "|" + strings.TrimSpace(agentName) + "|" + strings.TrimSpace(prompt) + "|" + time.Now().UTC().Format(time.RFC3339Nano)
+	return value
+}
+
+func runSummaryFailed(run *agentcomposev2.RunSummary) bool {
+	switch run.GetStatus() {
+	case agentcomposev2.RunStatus_RUN_STATUS_FAILED, agentcomposev2.RunStatus_RUN_STATUS_CANCELED:
+		return true
+	default:
+		return false
+	}
+}
+
+func runSummaryTerminal(run *agentcomposev2.RunSummary) bool {
+	switch run.GetStatus() {
+	case agentcomposev2.RunStatus_RUN_STATUS_SUCCEEDED, agentcomposev2.RunStatus_RUN_STATUS_FAILED, agentcomposev2.RunStatus_RUN_STATUS_CANCELED:
+		return true
+	default:
+		return false
+	}
+}
+
+func runSummaryExitCode(run *agentcomposev2.RunSummary) int {
+	if code := int(run.GetExitCode()); code > 0 && code < 126 {
+		return code
+	}
+	return exitCodeGeneral
+}
+
+func execResultExitCode(result *agentcomposev2.ExecResult) int {
+	if code := int(result.GetExitCode()); code > 0 && code < 126 {
+		return code
+	}
+	return exitCodeGeneral
+}
+
+func parseImagePlatform(value string) (*agentcomposev2.ImagePlatform, error) {
+	value = strings.TrimSpace(value)
+	if value == "" {
+		return nil, nil
+	}
+	parts := strings.Split(value, "/")
+	if len(parts) < 2 || len(parts) > 3 || strings.TrimSpace(parts[0]) == "" || strings.TrimSpace(parts[1]) == "" {
+		return nil, fmt.Errorf("invalid --platform %q: expected os/arch[/variant]", value)
+	}
+	platform := &agentcomposev2.ImagePlatform{
+		Os:           strings.TrimSpace(parts[0]),
+		Architecture: strings.TrimSpace(parts[1]),
+	}
+	if len(parts) == 3 {
+		platform.Variant = strings.TrimSpace(parts[2])
+	}
+	return platform, nil
+}
+
+func imageStoreText(store agentcomposev2.ImageStoreKind) string {
+	switch store {
+	case agentcomposev2.ImageStoreKind_IMAGE_STORE_KIND_DOCKER_DAEMON:
+		return "docker"
+	case agentcomposev2.ImageStoreKind_IMAGE_STORE_KIND_OCI_CACHE:
+		return "oci-cache"
+	default:
+		return "unspecified"
+	}
+}
+
+func imageAvailabilityStatusText(status agentcomposev2.ImageAvailabilityStatus) string {
+	switch status {
+	case agentcomposev2.ImageAvailabilityStatus_IMAGE_AVAILABILITY_STATUS_AVAILABLE:
+		return "available"
+	case agentcomposev2.ImageAvailabilityStatus_IMAGE_AVAILABILITY_STATUS_MISSING:
+		return "missing"
+	case agentcomposev2.ImageAvailabilityStatus_IMAGE_AVAILABILITY_STATUS_ERROR:
+		return "error"
+	default:
+		return "unspecified"
+	}
+}
+
+func imageOperationStatusText(status agentcomposev2.ImageOperationStatus) string {
+	switch status {
+	case agentcomposev2.ImageOperationStatus_IMAGE_OPERATION_STATUS_SUCCEEDED:
+		return "succeeded"
+	case agentcomposev2.ImageOperationStatus_IMAGE_OPERATION_STATUS_FAILED:
+		return "failed"
+	default:
+		return "unspecified"
+	}
+}
+
+func imagePlatformText(platform *agentcomposev2.ImagePlatform) string {
+	if platform == nil {
+		return ""
+	}
+	parts := []string{strings.TrimSpace(platform.GetOs()), strings.TrimSpace(platform.GetArchitecture())}
+	if strings.TrimSpace(platform.GetVariant()) != "" {
+		parts = append(parts, strings.TrimSpace(platform.GetVariant()))
+	}
+	if parts[0] == "" || parts[1] == "" {
+		return strings.Trim(strings.Join(parts, "/"), "/")
+	}
+	return strings.Join(parts, "/")
+}
+
+func shortImageID(id string) string {
+	id = strings.TrimPrefix(strings.TrimSpace(id), "sha256:")
+	if len(id) > 12 {
+		return id[:12]
+	}
+	return id
+}
+
+func cloneStringMapForCLI(values map[string]string) map[string]string {
+	if len(values) == 0 {
+		return nil
+	}
+	cloned := make(map[string]string, len(values))
+	for key, value := range values {
+		cloned[key] = value
+	}
+	return cloned
+}
+
+func runStatusText(status agentcomposev2.RunStatus) string {
+	switch status {
+	case agentcomposev2.RunStatus_RUN_STATUS_PENDING:
+		return "pending"
+	case agentcomposev2.RunStatus_RUN_STATUS_RUNNING:
+		return "running"
+	case agentcomposev2.RunStatus_RUN_STATUS_SUCCEEDED:
+		return "succeeded"
+	case agentcomposev2.RunStatus_RUN_STATUS_FAILED:
+		return "failed"
+	case agentcomposev2.RunStatus_RUN_STATUS_CANCELED:
+		return "canceled"
+	default:
+		return "unspecified"
+	}
+}
+
+func runSourceText(source agentcomposev2.RunSource) string {
+	switch source {
+	case agentcomposev2.RunSource_RUN_SOURCE_MANUAL:
+		return "manual"
+	case agentcomposev2.RunSource_RUN_SOURCE_SCHEDULER:
+		return "scheduler"
+	case agentcomposev2.RunSource_RUN_SOURCE_API:
+		return "api"
+	default:
+		return "unspecified"
+	}
+}
+
+func firstNonEmptyString(values ...string) string {
+	for _, value := range values {
+		if strings.TrimSpace(value) != "" {
+			return value
+		}
+	}
+	return ""
+}
